# gonna-pay-backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateCostRequest:
    properties:
      category:
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        type: string
      costName:
        type: string
      groupId:
        type: string
      ownerPercentage:
        type: number
      totalValue:
        type: number
    required:
    - category
    - costName
    - totalValue
    type: object
...
  request.UpdateCostRequest:
    properties:
      category:
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        type: string
      costName:
        type: string
      ownerPercentage:
        type: number
      totalValue:
        type: number
    type: object
...
  response.CostCategoryBreakdownResponse:
    properties:
      category:
        type: string
      percentage:
        type: number
      total:
        type: number
    type: object
  response.CostDetailResponse:
    properties:
      category:
        type: string
      costName:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      groupName:
        type: string
      id:
        type: string
      ownerPercentage:
        type: number
      ownerValue:
        type: number
      splitCount:
        type: integer
      splits:
        items:
          $ref: '#/definitions/response.SplitResponse'
        type: array
      totalValue:
        type: number
      updatedAt:
        type: string
    type: object
  response.CostResponse:
    properties:
      category:
        type: string
      costName:
        type: string
      createdAt:
        type: string
      groupId:
        type: string
      groupName:
        type: string
      id:
        type: string
      ownerPercentage:
        type: number
      ownerValue:
        type: number
      splitCount:
        type: integer
      totalValue:
        type: number
    type: object
  response.CostStatsResponse:
    properties:
      byCategory:
        items:
          $ref: '#/definitions/response.CostCategoryBreakdownResponse'
        type: array
      inSplits:
        type: number
      solo:
        type: number
      thisMonth:
        type: number
    type: object
  response.CostsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.CostResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      stats:
        $ref: '#/definitions/response.CostStatsResponse'
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  response.SplitResponse:
    properties:
      contactId:
        type: string
      contactName:
        type: string
      id:
        type: string
      percentage:
        type: number
      value:
        type: number
    type: object
...
  rest_errors.Causes:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  rest_errors.RestErrors:
    properties:
      causes:
        items:
          $ref: '#/definitions/rest_errors.Causes'
        type: array
      code:
        type: integer
      error:
        type: string
      message:
        type: string
    type: object
host: localhost:3333
info:
  contact: {}
  description: API de gestão financeira e divisão de custos
  title: Gonna Pay API
  version: "1.0"
paths:
...
  /cost:
    post:
      consumes:
      - application/json
      description: Cria um novo custo. Se groupId for informado, o valor é dividido
        entre os membros do grupo. ownerPercentage é opcional — se omitido, a divisão
        é igualitária
      parameters:
      - description: Dados do custo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateCostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CostDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Grupo não encontrado
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Criar custo
      tags:
      - costs
  /cost/{id}:
    delete:
      description: Remove um custo. Apenas o dono pode deletar
      parameters:
      - description: ID do custo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Deletar custo
      tags:
      - costs
    get:
      description: Retorna um custo com os detalhes de splits. Apenas o dono pode
        acessar
      parameters:
      - description: ID do custo
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CostDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Custo não pertence ao usuário
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Buscar custo por ID
      tags:
      - costs
    patch:
      consumes:
      - application/json
      description: Atualiza os dados de um custo. Apenas o dono pode editar. Os splits
        são recalculados automaticamente
      parameters:
      - description: ID do custo
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CostDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Atualizar custo
      tags:
      - costs
  /costs:
    get:
      description: Retorna todos os custos do usuário autenticado (sem detalhes dos
        splits)
      parameters:
      - description: 'Número da página (padrão: 1)'
        in: query
        name: page
        type: integer
      - description: 'Itens por página (padrão: 20, máximo: 100)'
        in: query
        name: limit
        type: integer
      - description: Filtrar por categoria
        enum:
        - Dinner
        - Lunch
        - Entertainment
        - Travel
        - Others
        in: query
        name: category
        type: string
      - description: Filtrar por período
        enum:
        - month
        - 7days
        in: query
        name: period
        type: string
      - description: Filtrar por tipo
        enum:
        - solo
        - group
        in: query
        name: type
        type: string
      - description: Valor mínimo
        in: query
        name: minValue
        type: number
      - description: Valor máximo
        in: query
        name: maxValue
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CostsListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest_errors.RestErrors'
      security:
      - CookieAuth: []
      summary: Listar custos
      tags:
      - costs
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"